editanndo y eliminado

+++ src/wall.js
@@ -44,24 +44,39 @@ const guardar = () => {
 //Leer documentos
 const postPrivado = () => {
   const tabla = document.getElementById('tabla');
+  
   db.collection("users").where("public", "==", 'Privado').onSnapshot((querySnapshot) => {
     let contenido = '';
+    const userId = firebase.auth().currentUser;
+    console.log(userId.uid)
     querySnapshot.forEach((doc) => {
-      contenido += `
-    <div class="card publish-content" id="${doc.id}">
-      <div class="card-content">
-        <span class="card-title">${doc.data().name}</span>
-        <p class="caja-post">${doc.data().first}</p>
-      </div>
-      <div class="card-action">
-       <a class="waves-effect waves-light btn indigo accent-1" onclick="editar('${doc.id}','${doc.data().first}')"><i class="fas fa-pen"></i>Editar</a>
-       <a class="waves-effect waves-light btn indigo accent-1" onclick="eliminar('${doc.id}')"><i class="fas fa-trash-alt"></i>Eliminar</a>
-      </div>
-      <div class="col s12 m6 l4 valign-wrapper">
-       <button class="likes-button" type="button" onclick = "incLikes('${doc.id}', '${doc.data().likes}')" ><i class="fas fa-heart"></i> Like  <span class="likes"></span></button>
-      </div>
-    </div>
-      `
+      if(userId.uid === doc.data().uidUser) {
+        contenido += `
+        <div class="publish-content" id="${doc.id}">
+          <p>${doc.data().name}</p>
+          <textarea class="materialize-textarea" id="${'myPost'+doc.id}" disabled>${doc.data().first}</textarea>
+          <div class="col s12 m6 l4 valign-wrapper">
+           <a class="waves-effect waves-light btn indigo accent-1" onclick="editar('${doc.id}','${doc.data().first}')" id="${'btnEdit'+doc.id}"><i class="fas fa-pen"></i>Editar</a>
+           <a class="waves-effect waves-light btn indigo accent-1" onclick="eliminar('${doc.id}')"><i class="fas fa-trash-alt"></i>Eliminar</a>
+          </div>
+          <div class="col s12 m6 l4 valign-wrapper">
+           <button class="likes-button" type="button" onclick = "incLikes('${doc.id}', '${doc.data().likes}')" ><i class="fas fa-heart"></i> Like  <span class="likes"></span></button>
+          </div>
+        </div>
+          `;
+
+      } else {
+        contenido += `
+        <div class="publish-content" id="${doc.id}">
+          <p>${doc.data().name}</p>
+          <textarea class="materialize-textarea" id="${'myPost'+doc.id}" disabled>${doc.data().first}</textarea>
+          <div class="col s12 m6 l4 valign-wrapper">
+           <button class="likes-button" type="button" onclick = "incLikes('${doc.id}', '${doc.data().likes}')" ><i class="fas fa-heart"></i> Like  <span class="likes"></span></button>
+          </div>
+        </div>
+          `;
+      }
+     
     });
     tabla.innerHTML = contenido
   });
@@ -70,37 +85,39 @@ const postPrivado = () => {
 //Borrar documentos
 const eliminar = (id) => {
   console.log('elimando', id);
-  db.collection("users").doc(id).delete().then(() => {
-  }).catch((error) => {
-    console.error("Error removing document: ", error);
-  });
-  const boton = document.getElementById('btnPublicar');
-  boton.onclick = guardar
+  const alertEliminar = confirm("Se eliminara este post");
+  if(alertEliminar === true){
+    db.collection("users").doc(id).delete().then(() => {
+    }).catch((error) => {
+      console.error("Error removing document: ", error);
+    });
+  }
 }
 
 //Editar documentos
 const editar = (id, post) => {
   console.log('editando', id)
-  document.getElementById('post').value = post;
-  const boton = document.getElementById('btnPublicar');
+  const textA = document.getElementById('myPost'+id);
+  textA.disabled = false;
+  document.getElementById('myPost'+id).value = post;
+  const boton = document.getElementById('btnEdit'+id);
+
   boton.innerHTML = 'Guardar';
   boton.onclick = () => {
-    const usersRef = db.collection("users").doc(id);
-    // Set the "capital" field of the city 'DC'
-    let post = document.getElementById('post').value;
-
-    return usersRef.update({
-      first: post,
+  const usersRef = db.collection("users").doc(id);
+ 
+  let post = document.getElementById('myPost'+id).value;
+  return usersRef.update({
+    first: post,
     })
-      .then(() => {
-        boton.innerHTML = 'Publicar';
-        document.getElementById('post').value = '';
-        boton.onclick = guardar
-      })
-      .catch((error) => {
-        // The document probably doesn't exist.
-        console.error("Error updating document: ", error);
-      });
+  .then(() => {
+  boton.innerHTML = 'Editar';
+  boton.onclick = editar;
+  })
+  .catch((error) => {
+   // The document probably doesn't exist.
+  console.error("Error updating document: ", error);
+  });
   }
 }
 
@@ -167,3 +184,4 @@ const incLikes = (id, likes) => {
     });
 
 }
+
